delete msctf_internal.h
